redux action and reducers folders pattern
