+++ src/store/actions/mainForm/index.ts
@@ -1,0 +1,6 @@
+import { ADD_DATA } from "store/types";
+
+export const mainFormData = (data: any) => ({
+  type: ADD_DATA,
+  payload: data,
+});
